Rename classes provided by the extension
