+++ ext/span.php
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-namespace OpenCensus\Trace;
+namespace OpenCensus\Trace\Ext;
 
 /**
  * This is the equivalent PHP class created by the opencensus C extension
